requirements.txt
Flask==2.1.1
werkzeug<2.2
langchain==0.0.1
faiss-cpu==1.7.1
neo4j==4.4.0
chroma==0.2.0
rasa==3.5.17
pandas==1.3.3
numpy==1.21.2
requests>=2.31,<3.0
PyPDF2==1.26.0
python-docx==0.8.11
openai==0.27.0
geopandas==0.9.0
whisper==1.0.0
spacy==3.7.2
blis==0.7.11
